+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Blender Export: stacked boards when using an automatically generated PCB3D.
+- Footprint variant: flipped components didn't get flipped after replacement
+  (#664)
 
 
 ## [1.8.0] - 2024-09-17